+++ common/src/main/java/fudge/notenoughcrashes/gui/InitErrorScreen.java
@@ -1,6 +1,6 @@
 package fudge.notenoughcrashes.gui;
 
-import fudge.notenoughcrashes.NecConfig;
+import fudge.notenoughcrashes.config.NecConfig;
 import fudge.notenoughcrashes.utils.NecLocalization;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
@@ -28,8 +28,6 @@ public class InitErrorScreen extends ProblemScreen {
 
         ButtonWidget exitButton = ButtonWidget.builder(NecLocalization.translatedText("menu.quit"),
                 button -> {
-                    // Prevent the game from freaking out when we try to close it
-                    NecConfig.instance().forceCrashScreen = false;
                     System.exit(-1);
                 }).dimensions(width / 2 - 155, height / 4 + 120 + 12, 150, 20).build();