+++ fabric/src/main/java/fudge/notenoughcrashes/fabric/FuckModrinth.java
@@ -1,4 +1,0 @@
-package fudge.notenoughcrashes.fabric;
-
-public class FuckModrinth {
-}